Change style names and split List.tsx into separate files
