+++ src/business/List/List.pcss
@@ -46,13 +46,15 @@
     }
   }
 
-  &--heading {
+  &__heading {
     padding: 1rem;
+    
+    &--compact {
+      padding: .75rem;
+    }
   }
 
-  &--compactHeading {
-    padding: .75rem;
-  }
+  
 
 
   &__decorator {
@@ -62,7 +64,7 @@
     text-align: center;
   }
 
-  &__mainContent {
+  &__content {
     flex: 0 1 100%;
     height: 2.75rem;
     display: flex;
